fix to make shader not crash on mac
Mac's implementation of OpenGL seems to crash (not compiler error,
TERRIBLE TERRIBLE RUNTIME CRASH) if functions are applied to globals in
their initialization outside of a function. This took hours to figure
out and a lot of pestering of friends with M1 Macs. I literally had to
binary search between a commit that i knew worked on mac, and the most
recent one to find the offending code. This sucked